requirements.txt
lxml>=4.2.1
pyyaml>=3.12
requests>=2.18.4